% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DualIndexedMatrix.R
\name{DualIndexedMatrix-class}
\alias{DualIndexedMatrix-class}
\alias{DualIndexedMatrix}
\alias{DualIndexedMatrixSeed}
\alias{DualIndexedMatrixSeed-class}
\title{Dual-indexed matrix representation}
\description{
Implements a dual-indexed \linkS4class{DelayedMatrix} representation that enables efficient row and column access at the cost of doubling the data.
}
\examples{
library(Matrix)
by.column <- rsparsematrix(100, 200, 0.1)
class(by.column) # dgCMatrices enable efficient column access.

by.row <- t(by.column) # the transposed columns are now the rows.
class(by.row)

mat <- DualIndexedMatrix(by.row, by.column, row.transposed=TRUE) 


}
\author{
Aaron Lun
}
